fix: defer fixer owner clear until after reviewer lock
Move resume-path ClearFixerOwnerToken past successful PR lock
reacquisition, make remove-failure tests deterministic without chmod,
and split dirty-adopt tests under the 300-line subsystem threshold.

Generated-By: looper 0.11.0 (runner=fixer, agent=grok-build)
