+++ internal/fixer/runner_dirty_adopt_test.go
@@ -5,8 +5,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/nexu-io/looper/internal/domain"
-	"github.com/nexu-io/looper/internal/loops"
 	"github.com/nexu-io/looper/internal/storage"
 )
 
@@ -143,248 +141,5 @@ func TestRunPrepareWorktreeStepRejectsDirtyWhenPathMatchesButOwnerTokenMissing(t
 	assertPrepareDirtyManualIntervention(t, checkpoint, err)
 	if len(git.inspectCalls) != 0 {
 		t.Fatalf("len(git.inspectCalls) = %d, want 0", len(git.inspectCalls))
-	}
-}
-
-func TestPreservedWorktreeOwnershipForRediscoveryCarriesUnpreparedOwnerToken(t *testing.T) {
-	t.Parallel()
-
-	path := filepath.Join(t.TempDir(), "wt-42")
-	token := "fixer:loop_1:run_1:t"
-	got := preservedWorktreeOwnershipForRediscovery(fixerCheckpoint{
-		Worktree: &checkpointWorktree{Path: path, Branch: "feature/fix-42", OwnerToken: token},
-	}, stepPrepareWorktree)
-	if got == nil || got.Path != path || got.Branch != "feature/fix-42" || got.OwnerToken != token || got.PreparedAt != "" {
-		t.Fatalf("preserved = %#v, want unprepared path+branch+token", got)
-	}
-
-	if preservedWorktreeOwnershipForRediscovery(fixerCheckpoint{
-		Worktree: &checkpointWorktree{Path: path, Branch: "feature/fix-42", OwnerToken: token, PreparedAt: "ready"},
-	}, stepPrepareWorktree) != nil {
-		t.Fatal("prepared worktree must not be carried for rediscovery ownership")
-	}
-	if preservedWorktreeOwnershipForRediscovery(fixerCheckpoint{
-		Worktree: &checkpointWorktree{Path: path, Branch: "feature/fix-42", OwnerToken: token},
-	}, stepRepair) != nil {
-		t.Fatal("non-prepare failures must not carry rediscovery ownership")
-	}
-	if preservedWorktreeOwnershipForRediscovery(fixerCheckpoint{
-		Worktree: &checkpointWorktree{Path: path, Branch: "feature/fix-42"},
-	}, stepPrepareWorktree) != nil {
-		t.Fatal("markerless path must not be carried as ownership")
-	}
-}
-
-func TestCreateRunContextPreservesOwnerTokenAcrossPrepareProbeRediscovery(t *testing.T) {
-	t.Parallel()
-
-	// Multi-attempt contract: failed prepare-worktree restarts discover, but the
-	// unprepared ownership slice must survive so the next prepare can same-head adopt.
-	fixture := newRunnerFixture(t)
-	f := newDirtyAdoptFixture(t)
-	repo := "acme/looper"
-	prNumber := int64(42)
-	loopTarget := "pr:acme/looper:42"
-	nowISO := fixture.nowISO()
-	ownerToken := "fixer:loop_prepare_rediscover:run_1:" + nowISO
-	if err := fixture.repos.Loops.Upsert(context.Background(), storage.LoopRecord{
-		ID:         "loop_fixer_prepare_rediscover_owner",
-		Seq:        1,
-		ProjectID:  "project_1",
-		Type:       "fixer",
-		TargetType: "pull_request",
-		TargetID:   &loopTarget,
-		Repo:       &repo,
-		PRNumber:   &prNumber,
-		Status:     "queued",
-		CreatedAt:  nowISO,
-		UpdatedAt:  nowISO,
-	}); err != nil {
-		t.Fatalf("Loops.Upsert() error = %v", err)
-	}
-	checkpointJSON := mustMarshalJSON(fixerCheckpoint{
-		Detail:   &checkpointDetail{HeadSHA: f.headSHA, HeadRefName: f.branch, BaseRefName: "main", State: "OPEN"},
-		FixItems: []FixItem{{Type: "comment", ID: "c1", ThreadID: "t1", Summary: "please fix"}},
-		// Prepare probe failed after rewind: path + ownership kept, PreparedAt empty.
-		Worktree: &checkpointWorktree{Path: f.wtPath, Branch: f.branch, OwnerToken: ownerToken},
-	})
-	if err := fixture.repos.Runs.Upsert(context.Background(), storage.RunRecord{
-		ID:             "run_failed_prepare_probe",
-		LoopID:         "loop_fixer_prepare_rediscover_owner",
-		Status:         "failed",
-		CurrentStep:    stringPtr(string(stepPrepareWorktree)),
-		CheckpointJSON: &checkpointJSON,
-		Summary:        stringPtr("error: cannot run ssh: No such file or directory"),
-		ErrorMessage:   stringPtr("error: cannot run ssh: No such file or directory"),
-		StartedAt:      nowISO,
-		CreatedAt:      nowISO,
-		UpdatedAt:      nowISO,
-	}); err != nil {
-		t.Fatalf("Runs.Upsert() error = %v", err)
-	}
-
-	runner := New(Options{DB: fixture.coordinator.DB(), Repos: fixture.repos, Logger: fixture.logger, Now: fixture.now})
-	loop, err := fixture.repos.Loops.GetByID(context.Background(), "loop_fixer_prepare_rediscover_owner")
-	if err != nil || loop == nil {
-		t.Fatalf("Loops.GetByID() = (%#v, %v)", loop, err)
-	}
-	created, err := runner.createRunContext(context.Background(), *loop)
-	if err != nil {
-		t.Fatalf("createRunContext() error = %v", err)
-	}
-	if created.StartStep != stepDiscoverPR {
-		t.Fatalf("StartStep = %q, want discover-pr", created.StartStep)
-	}
-	if created.Resumed {
-		t.Fatal("Resumed = true, want false for discover restart")
-	}
-	if created.Checkpoint.Worktree == nil {
-		t.Fatal("Worktree = nil, want ownership preserved across prepare rediscovery")
-	}
-	if created.Checkpoint.Worktree.Path != f.wtPath || created.Checkpoint.Worktree.OwnerToken != ownerToken {
-		t.Fatalf("Worktree = %#v, want path %q token %q", created.Checkpoint.Worktree, f.wtPath, ownerToken)
-	}
-	if created.Checkpoint.Worktree.PreparedAt != "" {
-		t.Fatalf("PreparedAt = %q, want empty", created.Checkpoint.Worktree.PreparedAt)
-	}
-	// Second prepare attempt with carried ownership must still same-head adopt.
-	f.seedOwnerToken(t, ownerToken)
-	git := &fakeGitGateway{
-		createResult:   CreateWorktreeResult{WorktreePath: f.wtPath, Branch: f.branch, HeadSHA: f.headSHA},
-		prepareResult:  PrepareWorktreeResult{HeadSHA: f.headSHA, Clean: false},
-		inspectResults: []InspectHeadResult{{HeadSHA: f.headSHA, HasUncommittedChanges: true}},
-	}
-	runner = New(Options{Git: git})
-	retryCheckpoint := created.Checkpoint
-	retryCheckpoint.Detail = &checkpointDetail{HeadSHA: f.headSHA, HeadRefName: f.branch, BaseRefName: "main", State: "OPEN"}
-	prepared, err := runner.runPrepareWorktreeStep(context.Background(), stepInput{
-		Project:    f.project(),
-		Loop:       storage.LoopRecord{ID: loop.ID, Status: "running"},
-		Run:        storage.RunRecord{ID: "run_retry"},
-		Repo:       repo,
-		PRNumber:   prNumber,
-		Checkpoint: retryCheckpoint,
-	})
-	if err != nil {
-		t.Fatalf("runPrepareWorktreeStep() after rediscovery error = %v", err)
-	}
-	if prepared.Worktree == nil || prepared.Worktree.PreparedAt == "" || prepared.Worktree.OwnerToken == "" {
-		t.Fatalf("prepared = %#v, want adopted worktree with new ownership stamp", prepared.Worktree)
-	}
-	if len(git.createCalls) != 0 {
-		t.Fatalf("len(git.createCalls) = %d, want 0 (reuse preserved path)", len(git.createCalls))
-	}
-}
-
-func TestRunPrepareWorktreeStepDirtyAdoptEarlyExitGatesSkipInspectHead(t *testing.T) {
-	t.Parallel()
-
-	takeoverMeta, err := loops.WriteTakeoverResume(nil, loops.TakeoverResume{SessionID: "sess-1"})
-	if err != nil {
-		t.Fatalf("WriteTakeoverResume() error = %v", err)
-	}
-
-	// Early-exit gates that must not call InspectHead, plus local mismatch which does.
-	// All cases carry fixer provenance so the gates under test are reached.
-	cases := []struct {
-		name        string
-		loop        storage.LoopRecord
-		detailHead  string
-		prepareHead string
-		inspectHead string // only used when inspectCalls want > 0
-		wantInspect int
-	}{
-		{
-			name:        "human_takeover",
-			loop:        storage.LoopRecord{ID: "loop_1", Status: string(domain.LoopStatusHumanTakeover)},
-			detailHead:  "base-head",
-			prepareHead: "base-head",
-			wantInspect: 0,
-		},
-		{
-			name:        "awaiting_human",
-			loop:        storage.LoopRecord{ID: "loop_1", Status: string(domain.LoopStatusAwaitingHuman)},
-			detailHead:  "base-head",
-			prepareHead: "base-head",
-			wantInspect: 0,
-		},
-		{
-			name:        "takeoverResume",
-			loop:        storage.LoopRecord{ID: "loop_1", Status: "running", MetadataJSON: &takeoverMeta},
-			detailHead:  "base-head",
-			prepareHead: "base-head",
-			wantInspect: 0,
-		},
-		{
-			name:        "empty_expected_head",
-			loop:        storage.LoopRecord{ID: "loop_1", Status: "running"},
-			detailHead:  "",
-			prepareHead: "base-head",
-			wantInspect: 0,
-		},
-		{
-			name:        "remote_mismatch",
-			loop:        storage.LoopRecord{ID: "loop_1", Status: "running"},
-			detailHead:  "base-head",
-			prepareHead: "remote-head",
-			wantInspect: 0,
-		},
-		{
-			name:        "local_mismatch",
-			loop:        storage.LoopRecord{ID: "loop_1", Status: "running"},
-			detailHead:  "base-head",
-			prepareHead: "base-head",
-			inspectHead: "other-head",
-			wantInspect: 1,
-		},
-	}
-
-	for _, tc := range cases {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			f := newDirtyAdoptFixture(t)
-			inspectResults := []InspectHeadResult{{HeadSHA: "base-head", HasUncommittedChanges: true}}
-			if tc.inspectHead != "" {
-				inspectResults = []InspectHeadResult{{HeadSHA: tc.inspectHead, HasUncommittedChanges: true}}
-			}
-			git := &fakeGitGateway{
-				createResult:   CreateWorktreeResult{WorktreePath: f.wtPath, Branch: f.branch, HeadSHA: f.headSHA},
-				prepareResult:  PrepareWorktreeResult{HeadSHA: tc.prepareHead, Clean: false},
-				inspectResults: inspectResults,
-			}
-			runner := New(Options{Git: git})
-			detail := &checkpointDetail{HeadRefName: f.branch, BaseRefName: "main"}
-			if tc.detailHead != "" {
-				detail.HeadSHA = tc.detailHead
-			}
-			ownerToken := "fixer:loop_1:run_prior:prepared"
-			f.seedOwnerToken(t, ownerToken)
-			checkpoint, err := runner.runPrepareWorktreeStep(context.Background(), stepInput{
-				Project:  f.project(),
-				Loop:     tc.loop,
-				Repo:     "acme/looper",
-				PRNumber: 42,
-				Checkpoint: fixerCheckpoint{
-					Detail: detail,
-					Worktree: &checkpointWorktree{
-						Path:       f.wtPath,
-						Branch:     f.branch,
-						OwnerToken: ownerToken,
-					},
-				},
-			})
-			assertPrepareDirtyManualIntervention(t, checkpoint, err)
-			if len(git.inspectCalls) != tc.wantInspect {
-				t.Fatalf("len(git.inspectCalls) = %d, want %d", len(git.inspectCalls), tc.wantInspect)
-			}
-			// Dirty non-adopt on rewind must not CleanupWorktree (preserve evidence).
-			if len(git.cleanupCalls) != 0 {
-				t.Fatalf("len(git.cleanupCalls) = %d, want 0 after rejected adopt", len(git.cleanupCalls))
-			}
-			if len(git.createCalls) != 0 {
-				t.Fatalf("len(git.createCalls) = %d, want 0 (no recreate after dirty MI)", len(git.createCalls))
-			}
-		})
 	}
 }
